backend/requirements.txt
flask
flask-cors 
googletrans==4.0.0-rc1 
google.generativeai
gunicorn
